# docs/swagger.yaml
basePath: /
definitions:
...
  code.Failure:
    properties:
      code:
        description: 业务码
        type: integer
      data:
        description: 业务返回数据
      message:
        description: 描述信息
        type: string
    type: object
...
  cron.detailResponse:
    properties:
      command:
        description: 执行命令
        type: string
      http_method:
        description: http 请求方式 1:get 2:post
        type: integer
      is_used:
        description: 是否启用 1:是  -1:否
        type: integer
      name:
        description: 任务名称
        type: string
      notify_keyword:
        description: 通知匹配关键字(多个用,分割)
        type: string
      notify_receiver_email:
        description: 通知者邮箱地址(多个用,分割)
        type: string
      notify_status:
        description: 执行结束是否通知 1:不通知 2:失败通知 3:结束通知 4:结果关键字匹配通知
        type: integer
      notify_type:
        description: 通知类型 1:邮件 2:webhook
        type: integer
      protocol:
        description: 执行方式 1:shell 2:http
        type: integer
      remark:
        description: 备注
        type: string
      retry_interval:
        description: 重试间隔(单位:秒)
        type: integer
      retry_times:
        description: 重试次数
        type: integer
      spec:
        description: crontab 表达式
        type: string
      timeout:
        description: 超时时间(单位:秒)
        type: integer
    type: object
...
  cron.modifyResponse:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
...
  menu.createActionResponse:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
  menu.createRequest:
    properties:
      icon:
        description: 图标
        type: string
      id:
        description: ID
        type: string
      level:
        description: 菜单类型 1:一级菜单 2:二级菜单
        type: integer
      link:
        description: 链接地址
        type: string
      name:
        description: 菜单名称
        type: string
      pid:
        description: 父类ID
        type: integer
    type: object
  menu.createResponse:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
...
  menu.deleteResponse:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
  menu.detailResponse:
    properties:
      icon:
        description: 图标
        type: string
      id:
        description: 主键ID
        type: integer
      link:
        description: 链接地址
        type: string
      name:
        description: 菜单名称
        type: string
      pid:
        description: 父类ID
        type: integer
    type: object
  menu.listActionData:
    properties:
      api:
        description: 调用方对接人
        type: string
      hash_id:
        description: hashID
        type: string
      menu_id:
        description: 菜单栏ID
        type: integer
      method:
        description: 调用方secret
        type: string
    type: object
  menu.listActionResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/menu.listActionData'
        type: array
      menu_name:
        type: string
    type: object
  menu.listData:
    properties:
      hashid:
        description: hashid
        type: string
      icon:
        description: 图标
        type: string
      id:
        description: ID
        type: integer
      is_used:
        description: 是否启用 1=启用 -1=禁用
        type: integer
      link:
        description: 链接地址
        type: string
      name:
        description: 菜单名称
        type: string
      pid:
        description: 父类ID
        type: integer
      sort:
        description: 排序
        type: integer
    type: object
  menu.listResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/menu.listData'
        type: array
    type: object
...
  menu.updateUsedResponse:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
...
info:
  contact: {}
  license:
    name: MIT
    url: https://github.com/ChangSZ/gin-boilerplate/blob/master/LICENSE
  title: swagger 接口文档
  version: "2.0"
paths:
...
  /api/cron/{id}:
    get:
      consumes:
      - application/json
      description: 获取单条任务详情
      parameters:
      - description: hashId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cron.detailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 获取单条任务详情
      tags:
      - API.cron
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 编辑任务
      parameters:
      - description: hashID
        in: formData
        name: id
        required: true
        type: string
      - description: 任务名称
        in: formData
        name: name
        required: true
        type: string
      - description: crontab 表达式
        in: formData
        name: spec
        required: true
        type: string
      - description: 执行命令
        in: formData
        name: command
        required: true
        type: string
      - description: 执行方式 1:shell 2:http
        in: formData
        name: protocol
        required: true
        type: integer
      - description: http 请求方式 1:get 2:post
        in: formData
        name: http_method
        type: integer
      - description: 超时时间(单位:秒)
        in: formData
        name: timeout
        required: true
        type: integer
      - description: 重试次数
        in: formData
        name: retry_times
        required: true
        type: integer
      - description: 重试间隔(单位:秒)
        in: formData
        name: retry_interval
        required: true
        type: integer
      - description: 执行结束是否通知 1:不通知 2:失败通知 3:结束通知 4:结果关键字匹配通知
        in: formData
        name: notify_status
        required: true
        type: integer
      - description: 通知类型 1:邮件 2:webhook
        in: formData
        name: notify_type
        type: integer
      - description: 通知者邮箱地址(多个用,分割)
        in: formData
        name: notify_receiver_email
        type: string
      - description: 通知匹配关键字(多个用,分割)
        in: formData
        name: notify_keyword
        type: string
      - description: 备注
        in: formData
        name: remark
        type: string
      - description: 是否启用 1:是  -1:否
        in: formData
        name: is_used
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cron.modifyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 编辑任务
      tags:
      - API.cron
...
  /api/menu:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 菜单列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/menu.listResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 菜单列表
      tags:
      - API.menu
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 创建/编辑菜单
      parameters:
      - description: 请求信息
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/menu.createRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/menu.createResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 创建/编辑菜单
      tags:
      - API.menu
  /api/menu/{id}:
    delete:
      consumes:
      - application/json
      description: 删除菜单
      parameters:
      - description: hashId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/menu.deleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 删除菜单
      tags:
      - API.menu
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 菜单详情
      parameters:
      - description: hashId
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/menu.detailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 菜单详情
      tags:
      - API.menu
...
  /api/menu/used:
    patch:
      consumes:
      - application/x-www-form-urlencoded
      description: 更新菜单为启用/禁用
      parameters:
      - description: hashId
        in: formData
        name: id
        required: true
        type: string
      - description: 是否启用 1:是 -1:否
        in: formData
        name: used
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/menu.updateUsedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 更新菜单为启用/禁用
      tags:
      - API.menu
  /api/menu_action:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 功能权限列表
      parameters:
      - description: hashID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/menu.listActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 功能权限列表
      tags:
      - API.menu
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 创建功能权限
      parameters:
      - description: HashID
        in: formData
        name: id
        required: true
        type: string
      - description: 请求方法
        in: formData
        name: method
        required: true
        type: string
      - description: 请求地址
        in: formData
        name: api
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/menu.createActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 创建功能权限
      tags:
      - API.menu
...
securityDefinitions:
  LoginToken:
    in: header
    name: token
    type: apiKey
swagger: "2.0"